requirements.txt
Django==2.2.9
django-filter==2.2.0
html5lib==1.0.1
Pillow==7.0.0
PyPDF2==1.26.0
pytz==2019.3
reportlab==3.5.34
six==1.14.0
sqlparse==0.3.0
webencodings==0.5.1
xhtml2pdf==0.2.4
